requirements.txt
# ============================================
# DevOps Butler - Dependencies
# ============================================

# --- Core Orchestration ---
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-aws>=0.2.0

# --- AWS SDK ---
boto3>=1.35.0
botocore>=1.35.0

# --- Code Parsing ---
tree-sitter>=0.23.0
tree-sitter-python>=0.23.0
tree-sitter-javascript>=0.23.0
tree-sitter-typescript>=0.23.0
tree-sitter-java>=0.23.0
tree-sitter-go>=0.23.0

# --- RAG / Embeddings ---
faiss-cpu>=1.8.0
numpy>=1.26.0

# --- Browser Automation ---
browser-use>=0.1.0
playwright>=1.45.0

# --- Web UI ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
python-multipart>=0.0.9

# --- CLI ---
click>=8.1.0

# --- Configuration ---
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0

# --- HTTP Client (HuggingFace API) ---
httpx>=0.27.0

# --- Utilities ---
pyyaml>=6.0
aiofiles>=24.1.0
